% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_S.R
\name{CMetafoR.ATE.S}
\alias{CMetafoR.ATE.S}
\title{Transporting ATE from multi-source population to an internal source-specific population}
\usage{
CMetafoR.ATE.S(
  X,
  Y,
  S,
  A,
  source_model = "SL.glmnet.multinom",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model = "SuperLearner",
  treatment_model_args = list(),
  outcome_model = "SuperLearner",
  outcome_model_args = list()
)
}
\arguments{
\item{X}{The covariate matrix/data frame with \eqn{n=n_1+...+n_m} rows and q coloums.}

\item{Y}{The (binary/categorical/continuous) outcome, which is a length \eqn{n} vector.}

\item{S}{The (numeric) source which is a length \eqn{n} vector.}

\item{A}{The (binary) treatment, which is a length \eqn{n} vector.}

\item{source_model}{The multi-nomial model for estimating \eqn{P(S=s|X)}. It has two options: \code{SL.glmnet.multinom} and \code{SL.nnet.multinom}. The default is \code{SL.glmnet.multinom}.}

\item{source_model_args}{The arguments (in \pkg{SuperLearner}) for the source model.}

\item{treatment_model_type}{The options for how the treatment_model \eqn{P(A=1|X, S=s)} is estimated. It includes \code{separate} and \code{joint}, with the default being \code{separate}. When \code{separate} is selected,
\eqn{P(A=1|X, S=s)} is estimated by fitting the model (regressing \eqn{A} on \eqn{X}) within each specific internal source population (S=s). When \code{joint} is selected, \eqn{P(A=1|X, S=s)}
is estimated by fitting the model (regressing \eqn{A} on \eqn{X} and \eqn{S}) using the multi-source population and then estimating the probability by fitting the model while suppressing the \eqn{S=s}.
In both cases, the propensity score is calculated as $P(A=1|X)=\sum_{s=1}^{m}P(A=1|X, S=s)P(S=s|X)$.}

\item{treatment_model}{The treatment model \eqn{P(A=1|X, S=s)} is estimated using \pkg{SuperLearner}. If, for example, the preference is to use only logistic regression for the probability estimation,
please ensure that only \code{glm} is included in the \pkg{SuperLearner} library within the \code{treatment_model_args}.}

\item{treatment_model_args}{The arguments (in \pkg{SuperLearner}) for the treatment model.}

\item{outcome_model}{The same as \code{treatment_model}.}

\item{outcome_model_args}{The arguments (in \pkg{SuperLearner}) for the outcome model.}
}
\value{

}
\description{
Doubly-robust and efficient estimator for the average treatment effects of each internal source-specific target population using \eqn{m} multi-source data.
}
\details{
Data structure: multi-source data contain outcome Y, source S, treatment A, and covariates X ($n \times p$).
The estimator is doubly robust and non-parametrically efficient. Three nuisance parameters are estimated,
the source model $q_s(X)=P(S=s|X)$, the propensity score model $\eta_a(X)=P(A=a|X)$, and the outcome model $\mu_a(X)=E(Y|X, A=a)$. The nuisance parameters are allowed to be estimated by \pkg{SuperLearner}. The estimator is
$$
 \dfrac{\widehat \kappa}{n}\sum\limits_{i=1}^{n} \Bigg[ I(S_i = s) \widehat \mu_a(X_i)
 +I(A_i = a) \dfrac{\widehat q_{s}(X_i)}{\widehat \eta_a(X_i)}  \Big\{ Y_i - \widehat \mu_a(X_i) \Big\} \Bigg],
$$
where $\widehat \kappa=\{n^{-1} \sum_{i=1}^n I(S_i=s)\}^{-1}$.
To achieve the non-parametrical efficiency and asymptotic normality, it requires that $||\widehat \mu_a(X) -\mu_a(X)||\big\{||\widehat \eta_a(X) -\eta_a(X)||+||\widehat q_s(X) -q_s(X)||\big\}=o_p(n^{-1/2})$.
In addition, to avoid the Donsker class assumption, the estimation is done by sample splitting and cross-fitting.
When one source of data is a randomized trial, it is still recommended to estimate the propensity score for optimal efficiency.
Since the non-parametric influence function is the same as the efficient semi-parametric efficient influence function when the propensity score is known and incorporating the assumption $Y\prep S|(X, A=a)$, the inference stays the same.
}
\references{
Dahabreh, I.J., Robertson, S.E., Petito, L.C., Hernán, M.A. and Steingrimsson, J.A.. (2019) \emph{Efficient and robust methods for causally
interpretable meta‐analysis: Transporting inferences from multiple randomized trials to a target population}, Biometrics.
}
