% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE_R.R
\name{CMetafoR.STE.R}
\alias{CMetafoR.STE.R}
\title{Transporting subgroup treatment effects (STE) from multi-source population to an external source-specific population}
\usage{
CMetafoR.STE.R(
  X,
  X0,
  Y,
  S,
  A,
  source_model = "SL.glmnet.multinom",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model = "SuperLearner",
  treatment_model_args = list(),
  R_model = "SuperLearner",
  R_model_args = list(),
  outcome_model = "SuperLearner",
  outcome_model_args = list(),
  x_tilde
)
}
\arguments{
\item{X}{The covariate matrix/data frame with \eqn{n=n_1+...+n_m} rows and q coloums. The first column of X is the categorical effect modifier ($\widetilde X$).}

\item{X0}{The covariate matrix/data frame with \eqn{n_0} rows and q coloums. The first column of X is the categorical effect modifier ($\widetilde X$).}

\item{Y}{The (binary/categorical/continuous) outcome, which is a length \eqn{n} vector.}

\item{S}{The (numeric) source which is a length \eqn{n} vector.}

\item{A}{The (binary) treatment, which is a length \eqn{n} vector.}

\item{source_model}{The multi-nomial model for estimating \eqn{P(S=s|X)}. It has two options: \code{SL.glmnet.multinom} and \code{SL.nnet.multinom}. The default is \code{SL.glmnet.multinom}.}

\item{source_model_args}{The arguments (in \pkg{SuperLearner}) for the source model.}

\item{treatment_model_type}{The options for how the treatment_model \eqn{P(A=1|X, S=s)} is estimated. It includes \code{separate} and \code{joint}, with the default being \code{separate}. When \code{separate} is selected, 
\eqn{P(A=1|X, S=s)} is estimated by fitting the model (regressing \eqn{A} on \eqn{X}) within each specific internal source population (S=s). When \code{joint} is selected, \eqn{P(A=1|X, S=s)} 
is estimated by fitting the model (regressing \eqn{A} on \eqn{X} and \eqn{S}) using the multi-source population and then estimating the probability by fitting the model while suppressing the \eqn{S=s}. 
In both cases, the propensity score is calculated as $P(A=1|X)=\sum_{s=1}^{m}P(A=1|X, S=s)P(S=s|X)$.}

\item{treatment_model}{The treatment model \eqn{P(A=1|X, S=s)} is estimated using \pkg{SuperLearner}. If, for example, the preference is to use only logistic regression for the probability estimation, 
please ensure that only \code{glm} is included in the \pkg{SuperLearner} library within the \code{treatment_model_args}.}

\item{treatment_model_args}{The arguments (in \pkg{SuperLearner}) for the treatment model.}

\item{R_model}{= The R model \eqn{P(R=1|W)} is estimated using \pkg{SuperLearner}. R is a binary variable indicating the multi-source data, i.e., R is 1 if the subject belongs to the multi-source data and 0 if the subject belongs to the external data.
W is combination of X and X0, i.e., W=rbind(X, X0)}

\item{R_model_args}{= list(),}

\item{outcome_model}{The same as \code{treatment_model}.}

\item{outcome_model_args}{The arguments (in \pkg{SuperLearner}) for the outcome model.}
}
\value{

}
\description{
Doubly-robust and efficient estimator for the subgroup treatment effects (STE) of an external target population using \eqn{m} multi-source data.
}
\details{
Data structure: multi-source data contain outcome Y, source S, treatment A, and covariates X ($n \times p$); external data contain only covariate X0 ($n0 \times p$). 
Once X and X0 are defined, The indicator of multi-source data, R, can be defined, i.e., R is 1 if the subject belongs to the multi-source data and 0 if the subject belongs to the external data. 
The estimator is doubly robust and non-parametrically efficient. Three nuisance parameters are estimated, 
the R model $q(X)=P(R=1|X)$, the propensity score model $\eta_a(X)=P(A=a|X)$, and the outcome model $\mu_a(X)=E(Y|X, A=a)$. The nuisance parameters are allowed to be estimated by \pkg{SuperLearner}. The estimator is
$$
 \dfrac{\widehat \kappa}{N}\sum\limits_{i=1}^{N} \Bigg[ I(R_i = 0) \widehat \mu_a(X_i) 
 +I(A_i = a, R_i=1) \dfrac{1-\widehat q(X_i)}{\widehat \eta_a(X_i)\widehat q(X_i)}  \Big\{ Y_i - \widehat \mu_a(X_i) \Big\} \Bigg],
$$
where $N=n+n_0$, and $\widehat \kappa=\{N^{-1} \sum_{i=1}^N I(R_i=0)\}^{-1}$.
To achieve the non-parametrical efficiency and asymptotic normality, it requires that $||\widehat \mu_a(X) -\mu_a(X)||\big\{||\widehat \eta_a(X) -\eta_a(X)||+||\widehat q(X) -q(X)||\big\}=o_p(n^{-1/2})$. 
In addition, to avoid the Donsker class assumption, the estimation is done by sample splitting and cross-fitting.
When one source of data is a randomized trial, it is still recommended to estimate the propensity score for optimal efficiency. 
Since the non-parametric influence function is the same as the efficient semi-parametric efficient influence function when the propensity score is known and incorporating the assumption $Y\prep S|(X, A=a)$, the inference stays the same.
}
