% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_external}
\alias{dat_external}
\title{External dataset}
\format{
A data frame with 10,083 rows and 13 columns. The columns are:
\tabular{ll}{
 \code{EM} \tab Effect modifier. \cr
 \code{X2}, ..., \code{X10}  \tab  Covariates. \cr
 \code{S} \tab Source indicator. \cr
 \code{A} \tab Treatment (1 for treated and 0 for untreated). \cr
 \code{Y} \tab Outcome.
}
}
\usage{
dat_external
}
\description{
Simulated external dataset.
}
\keyword{datasets}
