% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_nested}
\alias{dat_nested}
\title{Nested data}
\format{
`dat_nested`
A data frame with 1019 rows and 16 columns
\describe{
 \item{EM}{Effect modifier}
 \item{X2, ..., X10}{9 covariates}
 \item{R}{Indicator for nested data: R = 1 for nested data (not used)}
 \item{S}{Source}
 \item{A}{Treatment: 1 - treated; 0 - untreated}
 \item{Y}{Outcome}
 \item{Y1, Y0}{Counterfactual outcomes (not used)}
}
}
\usage{
dat_nested
}
\description{
Simulated nested dataset for CMetafoR
}
\keyword{datasets}
