% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.ATE_nested}
\alias{plot.ATE_nested}
\title{Plot method for objects of class "ATE_nested"}
\usage{
\method{plot}{ATE_nested}(x, source_names, ...)
}
\arguments{
\item{x}{Object of class "ATE_nested".}

\item{source_names}{optional, vector of character strings specifying the names of the sources. Defaults are the values in \code{S} provided by the user to \code{\link{ATE_nested}}.}

\item{...}{Other arguments, which are passed to \code{\link[metafor]{forest.rma}}.}
}
\value{
No value is returned.
}
\description{
This function creates forest plots of objects of class "ATE_nested".
}
\seealso{
\code{\link{ATE_nested}}
}
