% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.STE_int}
\alias{print.STE_int}
\title{Print method for objects of class "STE_int"}
\usage{
\method{print}{STE_int}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "STE_int".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "STE_int"
}
\seealso{
\code{\link{STE_int}}
}
