% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.STE_S}
\alias{print.STE_S}
\title{Print method for objects of class "STE_S"}
\usage{
\method{print}{STE_S}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "STE_S".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "STE_S"
}
\seealso{
\code{\link{CMetafoR.STE.S}}
}
