% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STE_int.R
\name{STE_int}
\alias{STE_int}
\title{Transporting subgroup treatment effects (STE) from multi-source population to an internal source-specific population}
\usage{
STE_int(
  X,
  Y,
  S,
  A,
  source_model = "glmnet.multinom",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model = "SuperLearner",
  treatment_model_args = list(),
  outcome_model = "SuperLearner",
  outcome_model_args = list()
)
}
\arguments{
\item{X}{The covariate data frame with \eqn{n=n_1+...+n_m} rows and q coloums. The first column of X is the categorical effect modifier (\eqn{\widetilde X}).}

\item{Y}{The (binary/categorical/continuous) outcome, which is a length \eqn{n} vector.}

\item{S}{The (numeric) source which is a length \eqn{n} vector.}

\item{A}{The (binary) treatment, which is a length \eqn{n} vector.}

\item{source_model}{The multi-nomial model for estimating \eqn{P(S=s|X)}. It has two options: \code{glmnet.multinom} and \code{nnet.multinom}. The default is \code{glmnet.multinom}.}

\item{source_model_args}{The arguments (in \pkg{SuperLearner}) for the source model.}

\item{treatment_model_type}{The options for how the treatment_model \eqn{P(A=1|X, S=s)} is estimated. It includes \code{separate} and \code{joint}, with the default being \code{separate}. When \code{separate} is selected,
\eqn{P(A=1|X, S=s)} is estimated by fitting the model (regressing \eqn{A} on \eqn{X}) within each specific internal source population (S=s). When \code{joint} is selected, \eqn{P(A=1|X, S=s)}
is estimated by fitting the model (regressing \eqn{A} on \eqn{X} and \eqn{S}) using the multi-source population and then estimating the probability by fitting the model while suppressing the \eqn{S=s}.
In both cases, the propensity score is calculated as \eqn{P(A=1|X)=\sum_{s=1}^{m}P(A=1|X, S=s)P(S=s|X)}.}

\item{treatment_model}{The treatment model \eqn{P(A=1|X, S=s)} is estimated using \pkg{SuperLearner}. If, for example, the preference is to use only logistic regression for the probability estimation,
please ensure that only \code{glm} is included in the \pkg{SuperLearner} library within the \code{treatment_model_args}.}

\item{treatment_model_args}{The arguments (in \pkg{SuperLearner}) for the treatment model.}

\item{outcome_model}{The same as \code{treatment_model}.}

\item{outcome_model_args}{The arguments (in \pkg{SuperLearner}) for the outcome model.}
}
\value{
An object of class "STE_int". This object is a list with the following elements:
  \item{df_dif}{A data frame containing the subgroup treatment effect (mean difference) estimates for the internal populations.}
  \item{df_A0}{A data frame containing the subgroup potential outcome mean estimates under A = 0 for the internal populations.}
  \item{df_A1}{A data frame containing the subgroup potential outcome mean estimates under A = 1 for the internal populations.}
  \item{fit_outcome}{Fitted outcome model.}
  \item{fit_source}{Fitted source model.}
  \item{fit_treatment}{Fitted treatment model(s).}
  \item{...}{Some additional elements.}
}
\description{
Doubly-robust and efficient estimator for the subgroup treatments effect (STE) of each internal source-specific target population using \eqn{m} multi-source data.
}
\details{
Data structure: multi-source data contain outcome Y, source S, treatment A, and covariates X (\eqn{n \times p}) with the first column being the categorical effect modifier (\eqn{\widetilde X}).
The estimator is doubly robust and non-parametrically efficient. Three nuisance parameters are estimated,
the source model \eqn{q_s(X)=P(S=s|X)}, the propensity score model \eqn{\eta_a(X)=P(A=a|X)}, and the outcome model \eqn{\mu_a(X)=E(Y|X, A=a)}. The nuisance parameters are allowed to be estimated by \pkg{SuperLearner}. The estimator is
\deqn{
 \dfrac{\widehat \kappa}{n}\sum\limits_{i=1}^{n} \Bigg[ I(S_i = s, \widetilde X_i=\widetilde x) \widehat \mu_a(X_i)
 +I(A_i = a, \widetilde X_i=\widetilde x) \dfrac{\widehat q_{s}(X_i)}{\widehat \eta_a(X_i)}  \Big\{ Y_i - \widehat \mu_a(X_i) \Big\} \Bigg],
}
where \eqn{\widehat \kappa=\{n^{-1} \sum_{i=1}^n I(S_i=s, \widetilde X_i=\widetilde x)\}^{-1}}.
To achieve the non-parametrical efficiency and asymptotic normality, it requires that \eqn{||\widehat \mu_a(X) -\mu_a(X)||\big\{||\widehat \eta_a(X) -\eta_a(X)||+||\widehat q_s(X) -q_s(X)||\big\}=o_p(n^{-1/2})}.
In addition, to avoid the Donsker class assumption, the estimation is done by sample splitting and cross-fitting.
When one source of data is a randomized trial, it is still recommended to estimate the propensity score for optimal efficiency.
Since the non-parametric influence function is the same as the efficient semi-parametric efficient influence function when the propensity score is known and incorporating the assumption \eqn{Y\perp S|(X, A=a)}, the inference stays the same.
}
