% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.STE_int}
\alias{summary.STE_int}
\title{Summary method for objects of class "STE_int"}
\usage{
\method{summary}{STE_int}(object, digits = 4, ...)
}
\arguments{
\item{object}{Object of class "STE_int".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Summary method for objects of class "STE_int"
}
\seealso{
\code{\link{STE_int}}
}
