% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.STE_internal}
\alias{summary.STE_internal}
\alias{summary.STE_external}
\alias{summary.ATE_external}
\alias{summary.ATE_internal}
\title{Summary method for objects of class "ATE_internal", "ATE_external", "STE_internal", or "STE_external"}
\usage{
\method{summary}{STE_internal}(object, digits = 4, ...)

\method{summary}{STE_external}(object, digits = 4, ...)

\method{summary}{ATE_external}(object, digits = 4, ...)

\method{summary}{ATE_internal}(object, digits = 4, ...)
}
\arguments{
\item{object}{Object of class "ATE_internal", "ATE_external", "STE_internal", or "STE_external".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Summary method for objects of class "ATE_internal", "ATE_external", "STE_internal", or "STE_external"
}
\examples{
si <- STE_internal(
  X = dat_multisource[, 2:10],
  Y = dat_multisource$Y,
  EM = dat_multisource$EM,
  S = dat_multisource$S,
  A = dat_multisource$A,
  cross_fitting = FALSE,
  source_model = "MN.nnet",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model_args = list(
    family = binomial(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  ),
  outcome_model_args = list(
    family = gaussian(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  )
)
summary(si)

}
\seealso{
\code{\link{ATE_internal}}, \code{\link{ATE_external}}, \code{\link{STE_internal}}, \code{\link{STE_external}}
}
