% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.STE_nested}
\alias{summary.STE_nested}
\alias{summary.STE_external}
\alias{summary.ATE_external}
\alias{summary.ATE_nested}
\title{Summary method for objects of class "ATE_nested", "ATE_external", "STE_nested", or "STE_external"}
\usage{
\method{summary}{STE_nested}(object, digits = 4, ...)

\method{summary}{STE_external}(object, digits = 4, ...)

\method{summary}{ATE_external}(object, digits = 4, ...)

\method{summary}{ATE_nested}(object, digits = 4, ...)
}
\arguments{
\item{object}{Object of class "ATE_nested", "ATE_external", "STE_nested", or "STE_external".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments.}
}
\value{
No value is returned.
}
\description{
Summary method for objects of class "ATE_nested", "ATE_external", "STE_nested", or "STE_external"
}
\seealso{
\code{\link{ATE_nested}}, \code{\link{ATE_external}}, \code{\link{STE_nested}}, \code{\link{STE_external}}
}
