% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_external.R
\name{ATE_external}
\alias{ATE_external}
\title{Transporting ATE from multi-source population to an external source-specific population}
\usage{
ATE_external(
  X,
  X_external,
  Y,
  S,
  A,
  cross_fitting = FALSE,
  replications = 10L,
  source_model = "MN.glmnet",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model_args = list(),
  external_model_args = list(),
  outcome_model_args = list(),
  show_progress = TRUE
)
}
\arguments{
\item{X}{The covariate data frame with \eqn{n=n_1+...+n_{|S|}} rows and \eqn{p} columns. Character variables will be converted to factors. Numeric variables will be used as is.}

\item{X_external}{The external covariate data frame with \eqn{n_0} rows and \eqn{p} columns. The external data counterpart to \code{X}.}

\item{Y}{The length \eqn{n} outcome vector.}

\item{S}{The source indicator which is a length \eqn{n} vector or factor. If \code{S} is a factor, it will maintain its level order, otherwise it will be converted to a factor with default level order. The order will be carried over to the outputs and plots.}

\item{A}{The binary treatment (1 for treated and 0 for untreated), which is a length \eqn{n} vector.}

\item{cross_fitting}{Logical, indicating whether sample splitting and cross fitting procedure should be used.}

\item{replications}{Integer, the number of sample splitting and cross fitting replications to performe, if \code{cross_fitting = TRUE}. Default is \code{10L}.}

\item{source_model}{The (penalized) multinomial logistic regression for estimating \eqn{P(S=s|X)}. It has two options: "\code{MN.glmnet}" (default) and "\code{MN.nnet}", which use \pkg{glmnet} and \pkg{nnet} respectively.}

\item{source_model_args}{The arguments (in \pkg{glmnet} or \pkg{nnet}) for the source model.}

\item{treatment_model_type}{How the propensity score \eqn{P(A=1|X)=\sum_{s \in S} P(A=1|X, S=s)P(S=s|X)} is estimated. Options include "\code{separate}" (default) and "\code{joint}". If "\code{separate}", \eqn{P(A=1|X, S=s)} is estimated by regressing \eqn{A} on \eqn{X} within each specific internal source population \eqn{S=s}. If "\code{joint}", \eqn{P(A=1|X, S=s)} is estimated by regressing \eqn{A} on \eqn{X} and \eqn{S} using the multi-source population.}

\item{treatment_model_args}{The arguments (in \pkg{SuperLearner}) for the treatment model.}

\item{external_model_args}{The arguments (in \pkg{SuperLearner}) for the external model.}

\item{outcome_model_args}{The arguments (in \pkg{SuperLearner}) for the outcome model.}

\item{show_progress}{Logical, indicating whether to print a progress bar for the cross-fit replicates completed, if \code{cross_fitting = TRUE}.}
}
\value{
An object of class "ATE_external". This object is a list with the following elements:
  \item{df_dif}{A data frame containing the treatment effect (mean difference) estimates for the extenal data.}
  \item{df_A0}{A data frame containing the potential outcome mean estimates under A = 0 for the extenal data.}
  \item{df_A1}{A data frame containing the potential outcome mean estimates under A = 1 for the extenal data.}
  \item{fit_outcome}{Fitted outcome model.}
  \item{fit_source}{Fitted source model.}
  \item{fit_treatment}{Fitted treatment model(s).}
  \item{fit_external}{Fitted external model.}
}
\description{
Doubly-robust and efficient estimator for the average treatment effect of an external target population using \eqn{m} multi-source data.
}
\details{
Data structure: multi-source data contain outcome Y, source S, treatment A, and covariates X (\eqn{n \times p}); external data contain only covariate X_external (\eqn{n_0 \times p}).
Once X and X_external are defined, The indicator of multi-source data, R, can be defined, i.e., R is 1 if the subject belongs to the multi-source data and 0 if the subject belongs to the external data.
The estimator is doubly robust and non-parametrically efficient. Three nuisance parameters are estimated,
the R model \eqn{q(X)=P(R=1|X)}, the propensity score model \eqn{\eta_a(X)=P(A=a|X)}, and the outcome model \eqn{\mu_a(X)=E(Y|X, A=a)}. The nuisance parameters are allowed to be estimated by \pkg{SuperLearner}. The estimator is
\deqn{
 \dfrac{\widehat \kappa}{N}\sum\limits_{i=1}^{N} \Bigg[ I(R_i = 0) \widehat \mu_a(X_i)
 +I(A_i = a, R_i=1) \dfrac{1-\widehat q(X_i)}{\widehat \eta_a(X_i)\widehat q(X_i)}  \Big\{ Y_i - \widehat \mu_a(X_i) \Big\} \Bigg],
}
where \eqn{N=n+n_0}, and \eqn{\widehat \kappa=\{N^{-1} \sum_{i=1}^N I(R_i=0)\}^{-1}}.
To achieve the non-parametrical efficiency and asymptotic normality, it requires that \eqn{||\widehat \mu_a(X) -\mu_a(X)||\big\{||\widehat \eta_a(X) -\eta_a(X)||+||\widehat q(X) -q(X)||\big\}=o_p(n^{-1/2})}.
In addition, to avoid the Donsker class assumption, the estimation is done by sample splitting and cross-fitting.
When one source of data is a randomized trial, it is still recommended to estimate the propensity score for optimal efficiency.
Since the non-parametric influence function is the same as the efficient semi-parametric efficient influence function when the propensity score is known and incorporating the assumption \eqn{Y\perp S|(X, A=a)}, the inference stays the same.
}
\examples{
ae <- ATE_external(
  X = dat_nested[, 2:10],
  Y = dat_nested$Y,
  S = dat_nested$S,
  A = dat_nested$A,
  X_external = dat_external[, 2:10],
  source_model = "MN.glmnet",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model_args = list(
    family = binomial(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  ),
  external_model_args = list(
    family = binomial(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  ),
  outcome_model_args = list(
    family = gaussian(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  )
)
}
\references{
Dahabreh, I.J., Robertson, S.E., Petito, L.C., Hernán, M.A. and Steingrimsson, J.A.. (2019) \emph{Efficient and robust methods for causally
interpretable meta‐analysis: Transporting inferences from multiple randomized trials to a target population}, Biometrics.
}
