% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_external}
\alias{dat_external}
\title{External data}
\format{
`dat_external`
A data frame with 8981 rows and 16 columns
\describe{
 \item{EM}{Effect modifier}
 \item{X2, ..., X10}{9 covariates}
 \item{R}{Indicator for nested data: R = 0 for external data (not used)}
 \item{S}{Source}
 \item{A}{Treatment: 1 - treated; 0 - untreated}
 \item{Y}{Outcome}
 \item{Y1, Y0}{Counterfactual outcomes (not used)}
}
}
\usage{
dat_external
}
\description{
Simulated external dataset for CMetafoR
}
\keyword{datasets}
