% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_multisource}
\alias{dat_multisource}
\title{Multi-source data}
\format{
`dat_multisource`
A data frame with 3917 rows and 13 columns
\describe{
 \item{EM}{Effect modifier}
 \item{X2, ..., X10}{9 covariates}
 \item{S}{Source}
 \item{A}{Treatment: 1 - treated; 0 - untreated}
 \item{Y}{Outcome}
}
}
\usage{
dat_multisource
}
\description{
Simulated multi-source dataset for CMetafoR
}
\keyword{datasets}
