% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.ATE_int}
\alias{plot.ATE_int}
\title{Plot method for objects of class "ATE_int"}
\usage{
\method{plot}{ATE_int}(x, header = "Study", xlab = "Treatment Effect", ...)
}
\arguments{
\item{x}{Object of class "ATE_int".}

\item{header}{character string specifying the title of the column of subgroup names}

\item{xlab}{character string specifying the title for the x-axis.}

\item{...}{Other arguments, which are passed to \code{\link[metafor]{forest.rma}}.}
}
\value{
No value is returned.
}
\description{
This function creates forest plots of objects of class "ATE_int".
}
\seealso{
\code{\link{ATE_int}}
}
