% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.ATE_nested}
\alias{plot.ATE_nested}
\title{Plot method for objects of class "ATE_nested"}
\usage{
\method{plot}{ATE_nested}(x, source_names, ...)
}
\arguments{
\item{x}{Object of class "ATE_nested".}

\item{source_names}{optional, vector of character strings specifying the names of the sources. Defaults are the values in \code{S} provided by the user to \code{\link{ATE_nested}}.}

\item{...}{Other arguments, which are passed to \code{\link[metafor]{forest.rma}}.}
}
\value{
No value is returned.
}
\description{
This function creates forest plots of objects of class "ATE_nested".
}
\examples{
an <- ATE_nested(
  X = dat_nested[, 1:10],
  Y = dat_nested$Y,
  S = dat_nested$S,
  A = dat_nested$A,
  source_model = "MN.glmnet",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model_args = list(
    family = binomial(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  ),
  outcome_model_args = list(
    family = gaussian(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  )
)
plot(an)

}
\seealso{
\code{\link{ATE_nested}}
}
