% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.STE_int}
\alias{plot.STE_int}
\title{Plot method for objects of class "STE_int"}
\usage{
\method{plot}{STE_int}(x, source_names, subgroup_names, use_scb = FALSE, ...)
}
\arguments{
\item{x}{Object of class "STE_int".}

\item{source_names}{vector of character strings specifying the names of the sources.}

\item{subgroup_names}{vector of character strings specifying the names of the subgroups.}

\item{use_scb}{logical scalar specifying whether the intervals in the forest plot should be simultaneous confidence bands (rather than confidence intervals). The default is \code{FALSE}.}

\item{...}{Other arguments, which are passed to \code{\link[metafor]{forest.rma}}.}
}
\value{
No value is returned.
}
\description{
This function creates forest plots of objects of class "STE_int".
}
\details{
Note that users may need to custom set the argument \code{ilab.xpos} which specifies the position (along the x-axis) of the subgroup labels. See \code{\link[metafor]{forest.rma}} for further details.
}
\seealso{
\code{\link{STE_int}}
}
