% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.STE_nested}
\alias{print.STE_nested}
\alias{print.ATE_nested}
\alias{print.STE_external}
\alias{print.ATE_external}
\title{Print method for objects of class "ATE_nested", "ATE_external", "STE_nested", or "STE_external"}
\usage{
\method{print}{STE_nested}(x, digits = 4, ...)

\method{print}{ATE_nested}(x, digits = 4, ...)

\method{print}{STE_external}(x, digits = 4, ...)

\method{print}{ATE_external}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "ATE_nested", "ATE_external", "STE_nested", or "STE_external".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments (ignored).}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "ATE_nested", "ATE_external", "STE_nested", or "STE_external"
}
\examples{
sn <- STE_nested(
  X = dat_multisource[, 2:10],
  Y = dat_multisource$Y,
  EM = dat_multisource$EM,
  S = dat_multisource$S,
  A = dat_multisource$A,
  cross_fitting = FALSE,
  source_model = "MN.nnet",
  source_model_args = list(),
  treatment_model_type = "separate",
  treatment_model_args = list(
    family = binomial(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  ),
  outcome_model_args = list(
    family = gaussian(),
    SL.library = c("SL.glmnet", "SL.nnet", "SL.glm"),
    cvControl = list(V = 5L)
  )
)
print(sn)

}
\seealso{
\code{\link{ATE_nested}}, \code{\link{ATE_external}}, \code{\link{STE_nested}}, \code{\link{STE_external}}
}
